move ProcessedAnnotationData to Base, with subclassed version for method which uses eppi-only attributes
